wip: create case helper (not courtnav)
