+++ e2e/tests/manageCases/caseProgression/checkApplication/checkApplication.spec.ts
@@ -1,6 +1,6 @@
 import { test } from "@playwright/test";
 import Config from "../../../../config";
-import createDaCitizenCourtNavCase from "../../../../common/createCaseHelper";
+import createDaCitizenCourtNavCase from "../../../../common/createDACourtNavCaseHelper";
 import { CheckApplication } from "../../../../journeys/manageCases/caseProgression/checkApplication/checkApplication";
 import { Helpers } from "../../../../common/helpers";
 import config from "../../../../config";
