+++ e2e/tests/manageCases/createCaseCCDPush.spec.ts
@@ -1,0 +1,6 @@
+import { test } from "@playwright/test";
+import createCaseSystemUser from "../../common/createCaseHelper";
+
+test("try create case", async () => {
+  await createCaseSystemUser();
+});
